Refactor/FriendsMain컴포넌트 분리작업
기존코드는 FriendsMain 하나의 컴포넌트내에서 status를 확인해 서버통신 및 전체 렌더링을 모두 처리하였으나 컴포넌트간 역할의 분리를 위해 리팩토링 작업 후 FriendsMain은 데이터를 schedule state에 저장하고 팔로우상태확인하여 렌더링여부 결정하는 역할만 하도록 변경됨.
